Separated SideMenu into several components
